requirements.txt
PySide6>=6.6.0
python-mpv>=1.0.6
PySide6-Frameless-Window>=0.4.3
pyinstaller>=6.0
watchdog>=6.0.0
